feat(fe): move mock data to own file
